move constant to graph
